refactor: #5
