+++ src/controllers/default-actions.controller.js
@@ -1,19 +1,19 @@
 const UserServices = require('blog-api-lib/services/UserServices');
-const { View } = require('../view/view');
+const { CommonView } = require('../view/common-views');
 const { Keyboards } = require('../keyboards/keyboard');
 const { RegisterView } = require('../view/register-view');
 class DefaultActionController {
 	#userServices;
 	#postServices;
 	#commentServices;
-	#view;
+	#commonView;
 	#registerView;
 	#keyboard;
 	constructor(UserServices, PostServices, CommentServices) {
 		this.#userServices = UserServices;
 		this.#postServices = PostServices;
 		this.#commentServices = CommentServices;
-		this.#view = new View(UserServices, PostServices, CommentServices);
+		this.#commonView = new CommonView(UserServices, PostServices, CommentServices);
 		this.#registerView = new RegisterView(UserServices, PostServices, CommentServices);
 		this.#keyboard = new Keyboards();
 	}
