+++ src/controllers/menus.controller.js
@@ -1,4 +1,4 @@
-const { View } = require('../view/view');
+const { CommonView } = require('../view/common-views');
 const { Keyboards } = require('../keyboards/keyboard');
 const { UserSettingView } = require('../view/user-setting-view');
 
@@ -6,14 +6,14 @@ class MenusController {
 	#userServices;
 	#postServices;
 	#commentServices;
-	#view;
+	#commonView;
 	#userSettingView;
 	#keyboard;
 	constructor(UserServices, PostServices, CommentServices) {
 		this.#userServices = UserServices;
 		this.#postServices = PostServices;
 		this.#commentServices = CommentServices;
-		this.#view = new View(UserServices, PostServices, CommentServices);
+		this.#commonView = new CommonView(UserServices, PostServices, CommentServices);
 		this.#userSettingView = new UserSettingView(UserServices, PostServices, CommentServices);
 		this.#keyboard = new Keyboards();
 	}
@@ -22,11 +22,11 @@ class MenusController {
 		if (ctx.message.text === ctx.i18n.t('buttons.userSetting')) {
 			this.#userSettingView.userSetting(ctx);
 		} else if (ctx.message.text === ctx.i18n.t('buttons.userPost')) {
-			this.#view.userPost(ctx);
+			this.#commonView.userPost(ctx);
 		} else if (ctx.message.text === ctx.i18n.t('buttons.getAnotherUser')) {
-			this.#view.getAnotherUser(ctx);
+			this.#commonView.getAnotherUser(ctx);
 		} else {
-			this.#view.mainMenu(ctx);
+			this.#commonView.mainMenu(ctx);
 		}
 	}
